requirements.txt
websockets>=12.0
python-dotenv>=1.0.0
MetaTrader5>=5.0.5572
sqlalchemy>=2.0.23
PyJWT>=2.8.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
xgboost>=2.0.3
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=2.1.3
